+++ native/headers/tx.h
@@ -4,6 +4,7 @@
 #include "shared.h"
 #include "lockable.h"
 #include "heap.h"
+#include "errors.h"
 
 // CONSTANTS
 #define MAX_TX (1 << 16)
@@ -23,8 +24,8 @@ typedef enum{
 
 typedef struct
 {
-    uint64_t old_heap_cursor;
-    uint64_t new_heap_cursor;
+    uint32_t old_heap_cursor;
+    uint32_t new_heap_cursor;
     uint64_t dep_tx_id;
     atomic_element_t* target;
 } TxOperation;
@@ -47,7 +48,7 @@ extern TxMap tx_map[MAX_TX];
 
 uint64_t      create_tx();
 Tx*           get_tx(uint64_t transaction_id);
-FnResponse    add_operation_tx(uint64_t old_cursor, uint64_t new_cursor, atomic_element_t* target, uint64_t tx_id, uint64_t dep_tx_id);
+FnResponse    add_operation_tx(uint32_t old_cursor, uint32_t new_cursor, atomic_element_t* target, uint64_t tx_id, uint64_t dep_tx_id);
 FnResponse    abort_operations_tx(TxOperation* operations, uint32_t operation_count);
 FnResponse    commit_tx(uint64_t transaction_id);
 FnResponse    abort_tx(uint64_t transaction_id);
