+++ native/src/core/tx/add_operation_tx.c
@@ -1,6 +1,6 @@
 #include "tx.h"
 
-FnResponse add_operation_tx(uint64_t old_cursor, uint64_t new_cursor, atomic_element_t* target, uint64_t tx_id, uint64_t dep_tx_id)
+FnResponse add_operation_tx(uint32_t old_cursor, uint32_t new_cursor, atomic_element_t* target, uint64_t tx_id, uint64_t dep_tx_id)
 {
 
     if ( tx_id <= dep_tx_id ) {
