+++ native/src/socle/heap/read_heap.c
@@ -1,5 +1,5 @@
 #include "heap.h"
 
-const heap_element* read_heap(uint64_t cursor){
+const heap_element* read_heap(uint32_t cursor){
     return (const heap_element*)&heap[cursor];
 }
